chore(listenerpolicy): clean up access log filter conversion some
